BugId:12003822:Fix some problems when using emulator on windows
Root Cause:
1> Cordova run yunos cann't launch emulator on windows since the avd
name is 'phone\r' which has a extra '\r' when list devices on windows
2> Resolving target failed since the format of the output of listing
devices is emu-xxx-id not emulator-id on windows
Solution:
Considering the difference between mac and windows, and modify the above
points to be compatible on windows.
Test Suggestoin:
Execute the command 'cordova run yunos' both on windows and mac, it can
launch the cordova app both on true device and emulator.

Change-Id: I12bf563655eb1a5354ecd56310f73009b9fb147a
Reviewed-on: http://gerrit3.alibaba-inc.com:8080/431459
Reviewed-by: 张婕欣 jessica.zjx <jessica.zjx@alibaba-inc.com>
Reviewed-by: 李广镇 guangzhen.lgz <guangzhen.lgz@alibaba-inc.com>